requirements.txt
# Airflow 2.9+ suporta Python 3.12
apache-airflow>=2.9.0
mlflow>=2.10.0

# Bibliotecas de Data Science (Versões ajustadas para existentes)
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.12.0

# Banco de Dados e Utilitários
sqlalchemy>=1.4,<2.0
psycopg2-binary>=2.9.0
joblib>=1.3.0
apache-airflow-providers-postgres>=5.0.0
